requirements.txt
Flask==2.3.2
requests==2.31.0
python-dotenv
supabase==3.6.1